# Move global caches to config class