--- a/.scalafmt.conf
+++ b/.scalafmt.conf
@@ -1,3 +1,4 @@
+version = "2.0.0-RC4"
 style = defaultWithAlign
 docstrings = JavaDoc
 assumeStandardLibraryStripMargin = true