--- a/nir/src/main/scala/scala/scalanative/nir/Global.scala
+++ b/nir/src/main/scala/scala/scalanative/nir/Global.scala
@@ -13,6 +13,7 @@ sealed abstract class Global {
   final def mangle: String =
     Mangle(this)
 }
+
 object Global {
   final case object None extends Global {
     override def top: Global.Top =