--- a/nir/src/main/scala/scala/scalanative/nir/serialization/BinaryDeserializer.scala
+++ b/nir/src/main/scala/scala/scalanative/nir/serialization/BinaryDeserializer.scala
@@ -76,7 +76,14 @@ final class BinaryDeserializer(buffer: ByteBuffer) {
   private def getString(): String = {
     val arr = new Array[Byte](getInt)
     get(arr)
-    new String(arr, "UTF-8")
+    val s = new String(arr, "UTF-8")
+    val ref = BinaryDeserializer.internedStrings.get(s)
+    if (ref == null) {
+      BinaryDeserializer.internedStrings.put(s, new java.lang.ref.WeakReference(s))
+      s
+    } else {
+      ref.get()
+    }
   }
 
   private def getBool(): Boolean = get != 0
@@ -306,4 +313,11 @@ final class BinaryDeserializer(buffer: ByteBuffer) {
     case T.StringVal  => Val.String(getString)
     case T.VirtualVal => Val.Virtual(getLong)
   }
+}
+
+object BinaryDeserializer {
+    import java.util.WeakHashMap
+    import java.lang.ref.WeakReference
+    private[scalanative] val internedStrings: WeakHashMap[String, WeakReference[String]] =
+      new WeakHashMap[String, WeakReference[String]]()
 }