The global caches have been moved to a cache called `SerializationCaches`, which is inside another cached called `RunCaches`. This last class intends to hold any caches that might be necessary during the run while the `SerializationCaches` contains only the caches used for serialization/deserialization purposes. Right now, only `deserializeBinary` uses them.