@jvican 2019-04-21
I've just run these optimizations in a stable machine for benchmarks.

Benchmarks before this PR:
  
```
> toolsBenchmarks/jmh:run -i 20 scala.scalanative.linker.benchmarks.HotTestSuiteLinkerBench
[info] Checking out Scala source version 2.11.12
[info] Compiling 2 Scala sources to /home/benchmarks/scala-native/tools-benchmarks/target/scala-2.10/classes...
Processing 6 classes from /home/benchmarks/scala-native/tools-benchmarks/target/scala-2.10/classes with "reflection" generator
Writing out Java source to /home/benchmarks/scala-native/tools-benchmarks/target/scala-2.10/src_managed/jmh and resources to /home/benchmarks/scala-native/tools-benchmarks/target/scala-2.10/resource_managed/jmh
[info] Compiling 1 Scala source and 5 Java sources to /home/benchmarks/scala-native/tools-benchmarks/target/scala-2.10/classes...
[info] Running org.openjdk.jmh.Main -i 20 scala.scalanative.linker.benchmarks.HotTestSuiteLinkerBench
[info] # JMH version: 1.21
[info] # VM version: JDK 1.8.0_181, Java HotSpot(TM) 64-Bit Server VM, 25.181-b13
[info] # VM invoker: /usr/lib/jvm/java-8-oracle/jre/bin/java
[info] # VM options: <none>
[info] # Warmup: 30 iterations, 500 ms each
[info] # Measurement: 20 iterations, 500 ms each
[info] # Timeout: 10 min per iteration
[info] # Threads: 1 thread, will synchronize iterations
[info] # Benchmark mode: Sampling time
[info] # Benchmark: scala.scalanative.linker.benchmarks.HotTestSuiteLinkerBench.linkTestSuite
[info] 
[info] # Run progress: 0.00% complete, ETA 00:00:25
[info] # Fork: 1 of 1
[info] # Warmup Iteration   1: [info] Linking (2805 ms)
[info] 2881.487 ms/op
[info] # Warmup Iteration   2: [info] Linking (1710 ms)
[info] 1709.179 ms/op
[info] # Warmup Iteration   3: [info] Linking (1120 ms)
[info] 1119.879 ms/op
[info] # Warmup Iteration   4: [info] Linking (967 ms)
[info] 967.836 ms/op
[info] # Warmup Iteration   5: [info] Linking (1005 ms)
[info] 1005.584 ms/op
[info] # Warmup Iteration   6: [info] Linking (919 ms)
[info] 918.553 ms/op
[info] # Warmup Iteration   7: [info] Linking (970 ms)
[info] 970.981 ms/op
[info] # Warmup Iteration   8: [info] Linking (979 ms)
[info] 979.370 ms/op
[info] # Warmup Iteration   9: [info] Linking (1193 ms)
[info] 1193.279 ms/op
[info] # Warmup Iteration  10: [info] Linking (1365 ms)
[info] 1365.246 ms/op
[info] # Warmup Iteration  11: [info] Linking (838 ms)
[info] 838.861 ms/op
[info] # Warmup Iteration  12: [info] Linking (846 ms)
[info] 846.201 ms/op
[info] # Warmup Iteration  13: [info] Linking (876 ms)
[info] 876.610 ms/op
[info] # Warmup Iteration  14: [info] Linking (827 ms)
[info] 827.326 ms/op
[info] # Warmup Iteration  15: [info] Linking (853 ms)
[info] 852.492 ms/op
[info] # Warmup Iteration  16: [info] Linking (905 ms)
[info] 904.921 ms/op
[info] # Warmup Iteration  17: [info] Linking (847 ms)
[info] 847.249 ms/op
[info] # Warmup Iteration  18: [info] Linking (892 ms)
[info] 892.338 ms/op
[info] # Warmup Iteration  19: [info] Linking (824 ms)
[info] 824.181 ms/op
[info] # Warmup Iteration  20: [info] Linking (848 ms)
[info] 848.298 ms/op
[info] # Warmup Iteration  21: [info] Linking (900 ms)
[info] 900.727 ms/op
[info] # Warmup Iteration  22: [info] Linking (844 ms)
[info] 844.104 ms/op
[info] # Warmup Iteration  23: [info] Linking (868 ms)
[info] 868.221 ms/op
[info] # Warmup Iteration  24: [info] Linking (824 ms)
[info] 824.181 ms/op
[info] # Warmup Iteration  25: [info] Linking (849 ms)
[info] 849.347 ms/op
[info] # Warmup Iteration  26: [info] Linking (901 ms)
[info] 900.727 ms/op
[info] # Warmup Iteration  27: [info] Linking (841 ms)
[info] 842.007 ms/op
[info] # Warmup Iteration  28: [info] Linking (886 ms)
[info] 886.047 ms/op
[info] # Warmup Iteration  29: [info] Linking (828 ms)
[info] 828.375 ms/op
[info] # Warmup Iteration  30: [info] Linking (850 ms)
[info] 850.395 ms/op
[info] Iteration   1: [info] Linking (896 ms)
[info] 896.532 ms/op
[info]                  linkTestSuite·p0.00:   896.532 ms/op
[info]                  linkTestSuite·p0.50:   896.532 ms/op
[info]                  linkTestSuite·p0.90:   896.532 ms/op
[info]                  linkTestSuite·p0.95:   896.532 ms/op
[info]                  linkTestSuite·p0.99:   896.532 ms/op
[info]                  linkTestSuite·p0.999:  896.532 ms/op
[info]                  linkTestSuite·p0.9999: 896.532 ms/op
[info]                  linkTestSuite·p1.00:   896.532 ms/op
[info] 
[info] Iteration   2: [info] Linking (839 ms)
[info] 838.861 ms/op
[info]                  linkTestSuite·p0.00:   838.861 ms/op
[info]                  linkTestSuite·p0.50:   838.861 ms/op
[info]                  linkTestSuite·p0.90:   838.861 ms/op
[info]                  linkTestSuite·p0.95:   838.861 ms/op
[info]                  linkTestSuite·p0.99:   838.861 ms/op
[info]                  linkTestSuite·p0.999:  838.861 ms/op
[info]                  linkTestSuite·p0.9999: 838.861 ms/op
[info]                  linkTestSuite·p1.00:   838.861 ms/op
[info] 
[info] Iteration   3: [info] Linking (890 ms)
[info] 890.241 ms/op
[info]                  linkTestSuite·p0.00:   890.241 ms/op
[info]                  linkTestSuite·p0.50:   890.241 ms/op
[info]                  linkTestSuite·p0.90:   890.241 ms/op
[info]                  linkTestSuite·p0.95:   890.241 ms/op
[info]                  linkTestSuite·p0.99:   890.241 ms/op
[info]                  linkTestSuite·p0.999:  890.241 ms/op
[info]                  linkTestSuite·p0.9999: 890.241 ms/op
[info]                  linkTestSuite·p1.00:   890.241 ms/op
[info] 
[info] Iteration   4: [info] Linking (821 ms)
[info] 821.035 ms/op
[info]                  linkTestSuite·p0.00:   821.035 ms/op
[info]                  linkTestSuite·p0.50:   821.035 ms/op
[info]                  linkTestSuite·p0.90:   821.035 ms/op
[info]                  linkTestSuite·p0.95:   821.035 ms/op
[info]                  linkTestSuite·p0.99:   821.035 ms/op
[info]                  linkTestSuite·p0.999:  821.035 ms/op
[info]                  linkTestSuite·p0.9999: 821.035 ms/op
[info]                  linkTestSuite·p1.00:   821.035 ms/op
[info] 
[info] Iteration   5: [info] Linking (850 ms)
[info] 850.395 ms/op
[info]                  linkTestSuite·p0.00:   850.395 ms/op
[info]                  linkTestSuite·p0.50:   850.395 ms/op
[info]                  linkTestSuite·p0.90:   850.395 ms/op
[info]                  linkTestSuite·p0.95:   850.395 ms/op
[info]                  linkTestSuite·p0.99:   850.395 ms/op
[info]                  linkTestSuite·p0.999:  850.395 ms/op
[info]                  linkTestSuite·p0.9999: 850.395 ms/op
[info]                  linkTestSuite·p1.00:   850.395 ms/op
[info] 
[info] Iteration   6: [info] Linking (902 ms)
[info] 901.775 ms/op
[info]                  linkTestSuite·p0.00:   901.775 ms/op
[info]                  linkTestSuite·p0.50:   901.775 ms/op
[info]                  linkTestSuite·p0.90:   901.775 ms/op
[info]                  linkTestSuite·p0.95:   901.775 ms/op
[info]                  linkTestSuite·p0.99:   901.775 ms/op
[info]                  linkTestSuite·p0.999:  901.775 ms/op
[info]                  linkTestSuite·p0.9999: 901.775 ms/op
[info]                  linkTestSuite·p1.00:   901.775 ms/op
[info] 
[info] Iteration   7: [info] Linking (851 ms)
[info] 851.444 ms/op
[info]                  linkTestSuite·p0.00:   851.444 ms/op
[info]                  linkTestSuite·p0.50:   851.444 ms/op
[info]                  linkTestSuite·p0.90:   851.444 ms/op
[info]                  linkTestSuite·p0.95:   851.444 ms/op
[info]                  linkTestSuite·p0.99:   851.444 ms/op
[info]                  linkTestSuite·p0.999:  851.444 ms/op
[info]                  linkTestSuite·p0.9999: 851.444 ms/op
[info]                  linkTestSuite·p1.00:   851.444 ms/op
[info] 
[info] Iteration   8: [info] Linking (890 ms)
[info] 890.241 ms/op
[info]                  linkTestSuite·p0.00:   890.241 ms/op
[info]                  linkTestSuite·p0.50:   890.241 ms/op
[info]                  linkTestSuite·p0.90:   890.241 ms/op
[info]                  linkTestSuite·p0.95:   890.241 ms/op
[info]                  linkTestSuite·p0.99:   890.241 ms/op
[info]                  linkTestSuite·p0.999:  890.241 ms/op
[info]                  linkTestSuite·p0.9999: 890.241 ms/op
[info]                  linkTestSuite·p1.00:   890.241 ms/op
[info] 
[info] Iteration   9: [info] Linking (831 ms)
[info] 831.521 ms/op
[info]                  linkTestSuite·p0.00:   831.521 ms/op
[info]                  linkTestSuite·p0.50:   831.521 ms/op
[info]                  linkTestSuite·p0.90:   831.521 ms/op
[info]                  linkTestSuite·p0.95:   831.521 ms/op
[info]                  linkTestSuite·p0.99:   831.521 ms/op
[info]                  linkTestSuite·p0.999:  831.521 ms/op
[info]                  linkTestSuite·p0.9999: 831.521 ms/op
[info]                  linkTestSuite·p1.00:   831.521 ms/op
[info] 
[info] Iteration  10: [info] Linking (853 ms)
[info] 853.541 ms/op
[info]                  linkTestSuite·p0.00:   853.541 ms/op
[info]                  linkTestSuite·p0.50:   853.541 ms/op
[info]                  linkTestSuite·p0.90:   853.541 ms/op
[info]                  linkTestSuite·p0.95:   853.541 ms/op
[info]                  linkTestSuite·p0.99:   853.541 ms/op
[info]                  linkTestSuite·p0.999:  853.541 ms/op
[info]                  linkTestSuite·p0.9999: 853.541 ms/op
[info]                  linkTestSuite·p1.00:   853.541 ms/op
[info] 
[info] Iteration  11: [info] Linking (901 ms)
[info] 900.727 ms/op
[info]                  linkTestSuite·p0.00:   900.727 ms/op
[info]                  linkTestSuite·p0.50:   900.727 ms/op
[info]                  linkTestSuite·p0.90:   900.727 ms/op
[info]                  linkTestSuite·p0.95:   900.727 ms/op
[info]                  linkTestSuite·p0.99:   900.727 ms/op
[info]                  linkTestSuite·p0.999:  900.727 ms/op
[info]                  linkTestSuite·p0.9999: 900.727 ms/op
[info]                  linkTestSuite·p1.00:   900.727 ms/op
[info] 
[info] Iteration  12: [info] Linking (851 ms)
[info] 851.444 ms/op
[info]                  linkTestSuite·p0.00:   851.444 ms/op
[info]                  linkTestSuite·p0.50:   851.444 ms/op
[info]                  linkTestSuite·p0.90:   851.444 ms/op
[info]                  linkTestSuite·p0.95:   851.444 ms/op
[info]                  linkTestSuite·p0.99:   851.444 ms/op
[info]                  linkTestSuite·p0.999:  851.444 ms/op
[info]                  linkTestSuite·p0.9999: 851.444 ms/op
[info]                  linkTestSuite·p1.00:   851.444 ms/op
[info] 
[info] Iteration  13: [info] Linking (875 ms)
[info] 875.561 ms/op
[info]                  linkTestSuite·p0.00:   875.561 ms/op
[info]                  linkTestSuite·p0.50:   875.561 ms/op
[info]                  linkTestSuite·p0.90:   875.561 ms/op
[info]                  linkTestSuite·p0.95:   875.561 ms/op
[info]                  linkTestSuite·p0.99:   875.561 ms/op
[info]                  linkTestSuite·p0.999:  875.561 ms/op
[info]                  linkTestSuite·p0.9999: 875.561 ms/op
[info]                  linkTestSuite·p1.00:   875.561 ms/op
[info] 
[info] Iteration  14: [info] Linking (823 ms)
[info] 823.132 ms/op
[info]                  linkTestSuite·p0.00:   823.132 ms/op
[info]                  linkTestSuite·p0.50:   823.132 ms/op
[info]                  linkTestSuite·p0.90:   823.132 ms/op
[info]                  linkTestSuite·p0.95:   823.132 ms/op
[info]                  linkTestSuite·p0.99:   823.132 ms/op
[info]                  linkTestSuite·p0.999:  823.132 ms/op
[info]                  linkTestSuite·p0.9999: 823.132 ms/op
[info]                  linkTestSuite·p1.00:   823.132 ms/op
[info] 
[info] Iteration  15: [info] Linking (853 ms)
[info] 852.492 ms/op
[info]                  linkTestSuite·p0.00:   852.492 ms/op
[info]                  linkTestSuite·p0.50:   852.492 ms/op
[info]                  linkTestSuite·p0.90:   852.492 ms/op
[info]                  linkTestSuite·p0.95:   852.492 ms/op
[info]                  linkTestSuite·p0.99:   852.492 ms/op
[info]                  linkTestSuite·p0.999:  852.492 ms/op
[info]                  linkTestSuite·p0.9999: 852.492 ms/op
[info]                  linkTestSuite·p1.00:   852.492 ms/op
[info] 
[info] Iteration  16: [info] Linking (898 ms)
[info] 898.630 ms/op
[info]                  linkTestSuite·p0.00:   898.630 ms/op
[info]                  linkTestSuite·p0.50:   898.630 ms/op
[info]                  linkTestSuite·p0.90:   898.630 ms/op
[info]                  linkTestSuite·p0.95:   898.630 ms/op
[info]                  linkTestSuite·p0.99:   898.630 ms/op
[info]                  linkTestSuite·p0.999:  898.630 ms/op
[info]                  linkTestSuite·p0.9999: 898.630 ms/op
[info]                  linkTestSuite·p1.00:   898.630 ms/op
[info] 
[info] Iteration  17: [info] Linking (848 ms)
[info] 848.298 ms/op
[info]                  linkTestSuite·p0.00:   848.298 ms/op
[info]                  linkTestSuite·p0.50:   848.298 ms/op
[info]                  linkTestSuite·p0.90:   848.298 ms/op
[info]                  linkTestSuite·p0.95:   848.298 ms/op
[info]                  linkTestSuite·p0.99:   848.298 ms/op
[info]                  linkTestSuite·p0.999:  848.298 ms/op
[info]                  linkTestSuite·p0.9999: 848.298 ms/op
[info]                  linkTestSuite·p1.00:   848.298 ms/op
[info] 
[info] Iteration  18: [info] Linking (897 ms)
[info] 896.532 ms/op
[info]                  linkTestSuite·p0.00:   896.532 ms/op
[info]                  linkTestSuite·p0.50:   896.532 ms/op
[info]                  linkTestSuite·p0.90:   896.532 ms/op
[info]                  linkTestSuite·p0.95:   896.532 ms/op
[info]                  linkTestSuite·p0.99:   896.532 ms/op
[info]                  linkTestSuite·p0.999:  896.532 ms/op
[info]                  linkTestSuite·p0.9999: 896.532 ms/op
[info]                  linkTestSuite·p1.00:   896.532 ms/op
[info] 
[info] Iteration  19: [info] Linking (829 ms)
[info] 829.424 ms/op
[info]                  linkTestSuite·p0.00:   829.424 ms/op
[info]                  linkTestSuite·p0.50:   829.424 ms/op
[info]                  linkTestSuite·p0.90:   829.424 ms/op
[info]                  linkTestSuite·p0.95:   829.424 ms/op
[info]                  linkTestSuite·p0.99:   829.424 ms/op
[info]                  linkTestSuite·p0.999:  829.424 ms/op
[info]                  linkTestSuite·p0.9999: 829.424 ms/op
[info]                  linkTestSuite·p1.00:   829.424 ms/op
[info] 
[info] Iteration  20: [info] Linking (849 ms)
[info] 849.347 ms/op
[info]                  linkTestSuite·p0.00:   849.347 ms/op
[info]                  linkTestSuite·p0.50:   849.347 ms/op
[info]                  linkTestSuite·p0.90:   849.347 ms/op
[info]                  linkTestSuite·p0.95:   849.347 ms/op
[info]                  linkTestSuite·p0.99:   849.347 ms/op
[info]                  linkTestSuite·p0.999:  849.347 ms/op
[info]                  linkTestSuite·p0.9999: 849.347 ms/op
[info]                  linkTestSuite·p1.00:   849.347 ms/op
[info] 
[info] 
[info] 
[info] Result "scala.scalanative.linker.benchmarks.HotTestSuiteLinkerBench.linkTestSuite":
[info]   N = 20
[info]   mean =    862.559 ±(99.9%) 24.531 ms/op
[info] 
[info]   Histogram, ms/op:
[info]     [820.000, 825.000) = 2 
[info]     [825.000, 830.000) = 1 
[info]     [830.000, 835.000) = 1 
[info]     [835.000, 840.000) = 1 
[info]     [840.000, 845.000) = 0 
[info]     [845.000, 850.000) = 2 
[info]     [850.000, 855.000) = 5 
[info]     [855.000, 860.000) = 0 
[info]     [860.000, 865.000) = 0 
[info]     [865.000, 870.000) = 0 
[info]     [870.000, 875.000) = 0 
[info]     [875.000, 880.000) = 1 
[info]     [880.000, 885.000) = 0 
[info]     [885.000, 890.000) = 0 
[info]     [890.000, 895.000) = 2 
[info]     [895.000, 900.000) = 3 
[info]     [900.000, 905.000) = 2 
[info] 
[info]   Percentiles, ms/op:
[info]       p(0.0000) =    821.035 ms/op
[info]      p(50.0000) =    851.968 ms/op
[info]      p(90.0000) =    900.517 ms/op
[info]      p(95.0000) =    901.723 ms/op
[info]      p(99.0000) =    901.775 ms/op
[info]      p(99.9000) =    901.775 ms/op
[info]      p(99.9900) =    901.775 ms/op
[info]      p(99.9990) =    901.775 ms/op
[info]      p(99.9999) =    901.775 ms/op
[info]     p(100.0000) =    901.775 ms/op
[info] 
[info] 
[info] # Run complete. Total time: 00:00:47
[info] 
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] 
[info] Benchmark                                                      Mode  Cnt    Score    Error  Units
[info] HotTestSuiteLinkerBench.linkTestSuite                        sample   20  862.559 ± 24.531  ms/op
[info] HotTestSuiteLinkerBench.linkTestSuite:linkTestSuite·p0.00    sample       821.035           ms/op
[info] HotTestSuiteLinkerBench.linkTestSuite:linkTestSuite·p0.50    sample       851.968           ms/op
[info] HotTestSuiteLinkerBench.linkTestSuite:linkTestSuite·p0.90    sample       900.517           ms/op
[info] HotTestSuiteLinkerBench.linkTestSuite:linkTestSuite·p0.95    sample       901.723           ms/op
[info] HotTestSuiteLinkerBench.linkTestSuite:linkTestSuite·p0.99    sample       901.775           ms/op
[info] HotTestSuiteLinkerBench.linkTestSuite:linkTestSuite·p0.999   sample       901.775           ms/op
[info] HotTestSuiteLinkerBench.linkTestSuite:linkTestSuite·p0.9999  sample       901.775           ms/op
[info] HotTestSuiteLinkerBench.linkTestSuite:linkTestSuite·p1.00    sample       901.775           ms/op
[success] Total time: 50 s, completed Apr 21, 2019 11:23:28 AM
```
  
Benchmarks after this PR:
  
```
[info] Running org.openjdk.jmh.Main -i 20 scala.scalanative.linker.benchmarks.HotTestSuiteLinkerBench
[info] # JMH version: 1.21
[info] # VM version: JDK 1.8.0_181, Java HotSpot(TM) 64-Bit Server VM, 25.181-b13
[info] # VM invoker: /usr/lib/jvm/java-8-oracle/jre/bin/java
[info] # VM options: <none>
[info] # Warmup: 30 iterations, 500 ms each
[info] # Measurement: 20 iterations, 500 ms each
[info] # Timeout: 10 min per iteration
[info] # Threads: 1 thread, will synchronize iterations
[info] # Benchmark mode: Sampling time
[info] # Benchmark: scala.scalanative.linker.benchmarks.HotTestSuiteLinkerBench.linkTestSuite
[info] 
[info] # Run progress: 0.00% complete, ETA 00:00:25
[info] # Fork: 1 of 1
[info] # Warmup Iteration   1: [info] Linking (1670 ms)
[info] 1746.928 ms/op
[info] # Warmup Iteration   2: [info] Linking (1577 ms)
[info] 1577.058 ms/op
[info] # Warmup Iteration   3: [info] Linking (1186 ms)
[info] 1186.988 ms/op
[info] # Warmup Iteration   4: [info] Linking (739 ms)
[info] 739.246 ms/op
[info] # Warmup Iteration   5: [info] Linking (611 ms)
[info] 611.320 ms/op
[info] # Warmup Iteration   6: [info] Linking (683 ms)
[info] 683.672 ms/op
[info] # Warmup Iteration   7: [info] Linking (709 ms)
[info] 708.837 ms/op
[info] # Warmup Iteration   8: [info] Linking (608 ms)
[info] 608.174 ms/op
[info] # Warmup Iteration   9: [info] Linking (685 ms)
[info] 685.769 ms/op
[info] # Warmup Iteration  10: [info] Linking (598 ms)
[info] 597.688 ms/op
[info] # Warmup Iteration  11: [info] Linking (647 ms)
[info] 646.971 ms/op
…